+++ src/lib/util.ts
@@ -6,7 +6,8 @@ import {
 	type Mod,
 	ModType,
 	ModLoader,
-	type ModId
+	type ModId,
+	type Prefs
 } from './types';
 import { convertFileSrc } from '@tauri-apps/api/core';
 import games from './state/game.svelte';
@@ -91,9 +92,9 @@ export function modIconSrc(mod: Mod) {
 
 export function gameIconSrc(game: Game) {
 	if (game.backends.length === 1 && game.backends[0] === Backend.Thunderstore) {
-		return `https://raw.githubusercontent.com/Kesomannen/gale/refs/heads/master/images/games/${game.slug}.webp`;
-	} else {
 		return `https://gcdn.thunderstore.io/assets/${game.slug}/${game.slug}-icon-192x192.webp`;
+	} else {
+		return `https://raw.githubusercontent.com/Kesomannen/gale/refs/heads/master/images/games/${game.slug}.webp`;
 	}
 }
 
@@ -193,10 +194,8 @@ export function loaderSupportsModpacks(loader: ModLoader) {
 	}
 }
 
-export async function shouldWarnForeginDownload(id: ModId): Promise<boolean> {
+export function shouldWarnForeignDownload(id: ModId, prefs: Prefs): boolean {
 	if (id.backend === Backend.Thunderstore) return false;
 	if (games.activeBackends.length === 1) return false;
-
-	const prefs = await api.prefs.get();
 	return !prefs.backendSkipConfirm;
 }