+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Button to view changelogs in the update mods dialog
+- Button to view changelogs in the update dialog
 - Support for Steam launch options (thanks [@hazre](https://github.com/hazre))
 
 ### Changed