+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Mod card lists such as the dependencies and hidden mods dialogs not being sorted correctly
 - Icons and links in update dialog always pointing to Thunderstore
 - Install button being stuck in a disabled state after installing mods
+- Sync socket failing to connect due to unsupported HTTP version
 
 ## 1.19.2 (2026-07-31)