+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Icons and links in update dialog always pointing to Thunderstore
 - Install button being stuck in a disabled state after installing mods
 - Sync socket failing to connect due to unsupported HTTP version
+- Minor UI fixes
 
 ## 1.19.2 (2026-07-31)